Minor example updates
Using `firmware_update_failed` in case statement with an `int` param now requires a cast to compile.

+++ src/assets/files/app-notes/AN028/firmware/src/Stop-Sleep-Cellular.cpp
@@ -286,7 +286,7 @@ void firmwareUpdateHandler(system_event_t event, int param) {
             break;
 
         case firmware_update_complete:
-        case firmware_update_failed:
+        case (int)firmware_update_failed:
             firmwareUpdateInProgress = false;
             break;
     }
